Move mock to outside pages folder to fix build error
